requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
openai==1.59.3
pydantic==2.10.4
pytest==8.3.4
httpx==0.28.1
python-dotenv==1.0.1
pytest-asyncio==0.25.2
